// apps/browser-app/src/components/Button/types.ts
export type ButtonColor =
  | 'white'
  | 'black'
  | 'light'
  | 'dark'
  | 'text'
  | 'ghost';

export type ButtonSize =
  | 'small'
  | 'normal'
  | 'medium'
  | 'large';

export type ButtonTheme =
  | 'light'
  | 'dark';

export type ButtonVariant =
  | 'primary'
  | 'link'
  | 'info'
  | 'success'
  | 'warning'
  | 'danger';